+++ src/main/java/com/gto/gtocore/api/playerskill/experiencelevel/NormalExperienceLevel.java
@@ -3,7 +3,6 @@ package com.gto.gtocore.api.playerskill.experiencelevel;
 import com.gto.gtocore.api.playerskill.experiencelevel.special.BodyExperienceLevel;
 
 public abstract class NormalExperienceLevel extends BasicExperienceLevel {
-
     protected BodyExperienceLevel bodyExperienceLevel;
 
     @Override
@@ -11,6 +10,7 @@ public abstract class NormalExperienceLevel extends BasicExperienceLevel {
         return (this.bodyExperienceLevel.level) << 1;
     }
 
+    @Override
     public void addExperience(int amount) {
         experience += amount;
         while (experience >= getExperienceForNextLevel() && level < getMaxLevel()) {
