+++ src/main/java/com/gto/gtocore/api/playerskill/logic/ExperienceSystemManager.java
@@ -3,7 +3,6 @@ package com.gto.gtocore.api.playerskill.logic;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.saveddata.SavedData;
-import net.minecraft.world.level.storage.DimensionDataStorage;
 
 import it.unimi.dsi.fastutil.objects.Object2LongMap;
 import it.unimi.dsi.fastutil.objects.Object2LongOpenHashMap;
